Update to breaking feature branch 1.0.0 (#73)
